% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_duncan.R
\name{scale_colour_duncan}
\alias{scale_colour_duncan}
\title{Duncan colours}
\usage{
scale_colour_duncan(...)
}
\description{
This function sets up the colours for duncans theme
scale_colour_duncan()
}
