% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_duncan.R
\name{theme_duncan}
\alias{theme_duncan}
\title{Duncan theme}
\usage{
theme_duncan(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\description{
This function sets up the theme for duncans theme
}
